+++ docs/docs/change_log.md
@@ -12,10 +12,13 @@ All notable changes to this project will be documented in this file.
 
 - Added tests and github workflow for automatic testing
 - Implemented _standarise_email() function
+- Added css styles
+- Add email preview to docs and README
 
 ### Fixed
 
 - Minor fixes in documentation
+- Fix number formatting
 - Add backwards compability with py-3.7 removing pandas kwargs that are not supported
 
 ## [0.1.0] - 2024-10-26